pyproject.toml
[project]
name = "drf-nested-file-serializaion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "djangorestframework>=3.15.2",
]
